Added files via upload
I'm using my DB currently in the code. When we merge with the project DB all that needs to be done is one line change in EventDAO.cs and adding a new table in the new DB. Here's the query.

CREATE TABLE [dbo].[event]
(
	[event_id] INT NOT NULL PRIMARY KEY IDENTITY, 
    [description] NVARCHAR(200) NULL, 
    [title] NVARCHAR(100) NULL, 
    [event_start] DATETIME NULL, 
    [event_end] DATETIME NULL, 
    [all_day] BIT NULL
)

I'm still working on putting a virtual keyboard for adding events via the touchscreen. Also tweaking the time start and time end code for adding events.